# Moved buffer extension into a separate method