**Before**

```
BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.25126.1000)
AMD Ryzen 7 4700U with Radeon Graphics, 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.300
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
```


|         Method |           Name |     Mean |   Error |  StdDev |
|--------------- |--------------- |---------:|--------:|--------:|
| SnakeLowerCase | XMLHttpRequest | 156.3 ns | 2.04 ns | 1.91 ns |
| SnakeUpperCase | XMLHttpRequest | 160.8 ns | 1.56 ns | 1.46 ns |
| KebabLowerCase | XMLHttpRequest | 155.8 ns | 1.53 ns | 1.43 ns |
| KebabUpperCase | XMLHttpRequest | 159.1 ns | 2.89 ns | 2.25 ns |

**After**

```
BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.25126.1000)
AMD Ryzen 7 4700U with Radeon Graphics, 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.300
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT AVX2
```


|         Method |           Name |     Mean |   Error |  StdDev |
|--------------- |--------------- |---------:|--------:|--------:|
| SnakeLowerCase | XMLHttpRequest | 145.6 ns | 0.85 ns | 0.75 ns |
| SnakeUpperCase | XMLHttpRequest | 148.5 ns | 0.94 ns | 0.83 ns |
| KebabLowerCase | XMLHttpRequest | 146.3 ns | 1.02 ns | 0.85 ns |
| KebabUpperCase | XMLHttpRequest | 147.1 ns | 0.71 ns | 0.67 ns |